+++ src/main/java/gov/milove/controllers/ExceptionHandlerController.java
@@ -4,6 +4,7 @@ import gov.milove.exceptions.NewsNotFoundException;
 import jakarta.persistence.EntityNotFoundException;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
