+++ src/main/java/gov/milove/controllers/forum/TopicController.java
@@ -3,7 +3,7 @@ package gov.milove.controllers.forum;
 import gov.milove.domain.dto.forum.NewTopicDto;
 import gov.milove.domain.dto.forum.TopicDto;
 import gov.milove.domain.forum.Topic;
-import gov.milove.domain.dto.document.media.TopicRepository;
+import gov.milove.repositories.jpa.media.TopicRepository;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
