+++ src/main/java/gov/milove/controllers/media/TextBannerController.java
@@ -1,7 +1,7 @@
 package gov.milove.controllers.media;
 
 import gov.milove.domain.media.TextBanner;
-import gov.milove.domain.dto.document.media.TextBannerRepository;
+import gov.milove.repositories.jpa.media.TextBannerRepository;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Example;
