+++ src/main/java/gov/milove/repositories/jpa/media/TopicRepository.java
@@ -1,4 +1,4 @@
-package gov.milove.domain.dto.document.media;
+package gov.milove.repositories.jpa.media;
 
 import gov.milove.domain.dto.forum.TopicDto;
 import gov.milove.domain.forum.Topic;
